localrequirements.txt
diff-match-patch==20121119
dj-static==0.0.6
Django==1.9.5
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-import-export==0.4.5
gunicorn==19.4.5
oauthlib==1.0.3
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
static3==0.6.1
tablib==0.11.2
stripe==1.32.2